Support for modules-as-images
This adds a FileRepo that will take Title's like "File:Module:Bananas.svg" and,
if a thumbnailed image does not exist, execute getMimeType on the Module, match
it with the extension (this means extensions need to match Mime-Types), and
then execute getImage on the Module, store it to a file and return it. It will
then be thumbnailed and shown as a normal image.
On errors, the error message will be generated in an SVG and the SVG will be
converted to the appropriate file type.
The generated files are scanned by the UploadBase verifier, so there are no XSS
issues that are not already present with normal SVG uploads.
Tests are almost just for coverage.

Change-Id: I14392b895cba36fbe84cebe58180c1ff5b6cea36